Added mapper in the menu

+++ src/main/java/com/mazzocchi/reservation/service/interfaces/IMenuService.java
@@ -6,6 +6,6 @@ import java.util.*;
 
 public interface IMenuService {
     List<Menu> findAllMenus();
-    Menu findMenuById(int id);
+    Menu findMenuById(Long id);
 
 }
